fix DI, product, cloudinary. deleted unnecessary dependencies

+++ DAL/Entities/Product.cs
@@ -12,8 +12,8 @@ namespace DAL.Entities
         public AvailableGenres Genre { get; set; }
         public AgeRating Rating { get; set; }
         public string Logo { get; set; }
-        public string Background { get; set; }
-        public int Price { get; set; }
+        public string Background { get; set; } 
+        public decimal Price { get; set; }
         public int Count { get; set; }
         public string DateCreated {  get; set; }
         public ICollection<ProductRating> Ratings {  get; set; }
